requirements.txt
Django==2.2.8
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
coverage==4.5.4
coveralls==1.9.2
factory-boy==2.12.0
django-cors-headers==3.2.1
django-extensions==2.2.6